@@ -32,55 +25,57 @@
     "email": "dbx-node@mongodb.com"
   },
   "dependencies": {
-    "bson": "^4.6.1",
+    "bson": "^4.6.2",
     "denque": "^2.0.1",
-    "mongodb-connection-string-url": "^2.4.1",
-    "socks": "^2.6.1"
+    "mongodb-connection-string-url": "^2.5.2",
+    "socks": "^2.6.2"
   },
   "devDependencies": {
+    "@iarna/toml": "^2.2.5",
     "@istanbuljs/nyc-config-typescript": "^1.0.2",
-    "@microsoft/api-extractor": "^7.19.4",
+    "@microsoft/api-extractor": "^7.20.0",
     "@microsoft/tsdoc-config": "^0.15.2",
     "@types/chai": "^4.3.0",
     "@types/chai-subset": "^1.3.3",
+    "@types/express": "^4.17.13",
     "@types/kerberos": "^1.1.1",
-    "@types/mocha": "^9.0.0",
-    "@types/node": "^17.0.10",
+    "@types/mocha": "^9.1.0",
+    "@types/node": "^17.0.23",
     "@types/saslprep": "^1.0.1",
     "@types/semver": "^7.3.9",
-    "@types/sinon": "^10.0.6",
+    "@types/sinon": "^10.0.11",
     "@types/whatwg-url": "^8.2.1",
-    "@typescript-eslint/eslint-plugin": "^5.10.0",
-    "@typescript-eslint/parser": "^5.10.0",
+    "@typescript-eslint/eslint-plugin": "^5.17.0",
+    "@typescript-eslint/parser": "^5.17.0",
     "bluebird": "^3.7.2",
-    "chai": "^4.3.4",
+    "chai": "^4.3.6",
     "chai-subset": "^1.6.0",
     "chalk": "^4.1.2",
-    "downlevel-dts": "^0.8.0",
-    "eslint": "^8.7.0",
-    "eslint-config-prettier": "^8.3.0",
+    "eslint": "^8.12.0",
+    "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-import": "^2.25.4",
     "eslint-plugin-prettier": "^4.0.0",
     "eslint-plugin-simple-import-sort": "^7.0.0",
     "eslint-plugin-tsdoc": "^0.2.14",
+    "express": "^4.17.3",
     "js-yaml": "^4.1.0",
     "lodash.camelcase": "^4.3.0",
-    "mocha": "^9.1.4",
+    "mocha": "^9.2.2",
     "mocha-sinon": "^2.1.2",
     "nyc": "^15.1.0",
-    "prettier": "^2.5.1",
+    "prettier": "^2.6.1",
     "rimraf": "^3.0.2",
     "semver": "^7.3.5",
-    "sinon": "^12.0.1",
+    "sinon": "^13.0.1",
     "sinon-chai": "^3.7.0",
     "source-map-support": "^0.5.21",
     "standard-version": "^9.3.2",
-    "ts-node": "^10.4.0",
-    "tsd": "^0.19.1",
-    "typescript": "^4.5.4",
+    "ts-node": "^10.7.0",
+    "tsd": "^0.20.0",
+    "typescript": "^4.6.3",
     "typescript-cached-transpile": "^0.0.6",
     "xml2js": "^0.4.23",
-    "yargs": "^17.3.1"
+    "yargs": "^17.4.0"
   },
   "license": "Apache-2.0",
   "engines": {
