@@ -1,7 +1,7 @@
 {
   "name": "mongoose",
   "description": "Mongoose MongoDB ODM",
-  "version": "6.2.9",
+  "version": "6.3.0",
   "author": "Guillermo Rauch <guillermo@learnboost.com>",
   "keywords": [
     "mongodb",
@@ -21,7 +21,7 @@
   "dependencies": {
     "bson": "^4.2.2",
     "kareem": "2.3.5",
-    "mongodb": "4.3.1",
+    "mongodb": "4.5.0",
     "mpath": "0.8.4",
     "mquery": "4.0.2",
     "ms": "2.1.3",
@@ -30,47 +30,53 @@
   "devDependencies": {
     "@babel/core": "7.17.8",
     "@babel/preset-env": "7.16.11",
-    "@typescript-eslint/eslint-plugin": "5.16.0",
-    "@typescript-eslint/parser": "5.16.0",
-    "acquit": "1.x",
-    "acquit-ignore": "0.2.x",
-    "acquit-require": "0.1.x",
+    "@typescript-eslint/eslint-plugin": "5.17.0",
+    "@typescript-eslint/parser": "5.17.0",
+    "acquit": "1.2.1",
+    "acquit-ignore": "0.2.0",
+    "acquit-require": "0.1.1",
+    "assert-browserify": "npm:assert@^2.0.0",
     "axios": "0.26.1",
     "babel-loader": "8.2.4",
     "benchmark": "2.1.4",
     "bluebird": "3.7.2",
     "cheerio": "1.0.0-rc.10",
+    "crypto-browserify": "3.12.0",
     "dox": "0.3.1",
-    "eslint": "8.11.0",
+    "eslint": "8.12.0",
     "eslint-plugin-mocha-no-only": "1.1.1",
-    "highlight.js": "9.18.3",
+    "highlight.js": "11.5.0",
     "lodash.isequal": "4.5.0",
     "lodash.isequalwith": "4.4.0",
     "marked": "4.0.12",
     "mocha": "9.2.2",
     "moment": "2.x",
-    "mongodb-memory-server": "^8.3.0",
-    "nyc": "^15.1.0",
+    "mongodb-memory-server": "8.3.0",
+    "nyc": "15.1.0",
     "pug": "3.0.2",
     "q": "1.5.1",
     "serve-handler": "6.1.3",
+    "sinon": "13.0.0",
+    "stream-browserify": "3.0.0",
     "tsd": "0.19.1",
     "typescript": "4.6.3",
     "uuid": "8.3.2",
-    "webpack": "4.44.1"
+    "webpack": "5.70.0"
   },
   "directories": {
     "lib": "./lib/mongoose"
   },
   "scripts": {
     "lint": "eslint .",
+    "lint-js": "eslint . --ext .js",
+    "lint-ts": "eslint . --ext .ts",
     "build-browser": "node build-browser.js",
     "prepublishOnly": "npm run build-browser",
     "release": "git pull && git push origin master --tags && npm publish",
     "release-legacy": "git pull origin 5.x && git push origin 5.x --tags && npm publish --tag legacy",
     "mongo": "node ./tools/repl.js",
     "test": "mocha --exit ./test/*.test.js",
-    "test-tsd": "tsd",
+    "test-tsd": "node ./test/types/check-types-filename && tsd",
     "tdd": "mocha ./test/*.test.js ./test/typescript/main.test.js --inspect --watch --recursive --watch-files ./**/*.js",
     "test-coverage": "nyc --reporter=html --reporter=text npm test"
   },
